refactor: エクスポートモジュールをlib/agg/からlib/export/に移動
集計ロジック(agg/)とエクスポート(export/)の責務を分離。
依存方向はexport/ → agg/の一方向。
